+++ semgrep_rules_check/eslint_rules/object_injection_negatives.js
@@ -1,0 +1,10 @@
+function foo(someVar, obj) {
+  two = 2
+  string = "key"
+
+  value = obj.whyNot
+  value = obj['1']
+  value = obj[two]
+  value = obj[string]
+  value = obj[7]
+}