+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Semgrep analyzer changelog
 - Add react-dangerouslysetinnerhtml semgrep rule (!16)
 - Add detect-non-literal-regexp semgrep rule (!16)
 - Add detect-non-literal-fs-filename (!16)
+- Add detect-object-injection (!16)
 
 ## v0.6.0
 - Add eslint and react rule-sets (!12)